Move ElidedLabel into its own file; cleanup
