+++ OrbitQt/orbitlivefunctions.ui
@@ -62,7 +62,7 @@
    <item row="2" column="0">
     <layout class="QVBoxLayout" name="iteratorLayout">
      <item>
-      <widget class="QLabel" name="label">
+      <widget class="QLabel" name="instructionsLabel">
        <property name="text">
         <string>Select functions above, right-click, and select 'Add iterator(s)' to add iterators for instrumented functions.</string>
        </property>
